requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==2.2
django-cors-headers==3.6.0
django-filter==2.4.0
django-nested-admin==3.3.3
django-templated-mail==1.1.1
django-tinymce==3.2.0
djangorestframework==3.12.2
djangorestframework-jsonp==1.0.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
Markdown==3.3.3
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.1.0
progress==1.5
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.6
python-decouple==3.4
python-monkey-business==1.0.0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
